requirements.txt
fastapi
uvicorn[standard]
transformers==4.41.2
Pillow
python-multipart
numpy<2

torch==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
